requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
aiohttp
feedparser
cloudscraper
curl_cffi
beautifulsoup4
requests
requests-toolbelt
setuptools
soupsieve
typing_extensions
urllib3
psycopg2-binary
lxml
google-genai
playwright
playwright-stealth
ccxt
yfinance
websocket
websocket-client